add claims to jwt

+++ resources/item.py
@@ -1,4 +1,4 @@
-import uuid
+from flask_jwt_extended import jwt_required, get_jwt
 from flask import Flask, request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
@@ -17,8 +17,13 @@ class Item(MethodView):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     @blp.response(200)
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege is required.")
+
         item = ItemModel.query.get_or_404(item_id)
         try:
             db.session.delete(item)
@@ -50,6 +55,7 @@ class ItemList(MethodView):
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
